+++ backend/src/main/java/middle_point_search/backend/domains/recommendPlace/controller/RecommendPlaceController.java
@@ -15,12 +15,13 @@ import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import middle_point_search.backend.common.dto.DataResponse;
 import middle_point_search.backend.common.dto.ErrorResponse;
-import middle_point_search.backend.domains.market.dto.request.RecommendPlacesFindRequest;
-import middle_point_search.backend.domains.market.dto.response.RecommendPlacesFindResponse;
+import middle_point_search.backend.domains.recommendPlace.dto.request.RecommendPlacesFindRequest;
+import middle_point_search.backend.domains.recommendPlace.dto.response.RecommendPlacesFindResponse;
 import middle_point_search.backend.domains.recommendPlace.service.RecommendPlaceService;
 
 @Tag(name = "RECOMMEND PLACE API",  description = "추천 장소에 대한 API입니다.")
@@ -75,7 +76,7 @@ public class RecommendPlaceController {
 		}
 	)
 	public ResponseEntity<DataResponse<Page<RecommendPlacesFindResponse>>> recommendPlacesFind(
-		@ModelAttribute @ParameterObject RecommendPlacesFindRequest request) {
+		@Valid @ModelAttribute @ParameterObject RecommendPlacesFindRequest request) {
 
 		Page<RecommendPlacesFindResponse> recommendPlaces = recommendPlaceService.findRecommendPlaces(request);
